requirements.txt
uvicorn~=0.32.0
fastapi~=0.115.3
sqlmodel~=0.0.22
redis~=5.2.0
python-dotenv~=1.0.1